requirements.txt
altgraph==0.17.4
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
customtkinter==5.2.2
darkdetect==0.8.0
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PySocks==1.7.1
requests==2.32.5
selenium==4.35.0
setuptools==80.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wheel==0.45.1
wsproto==1.2.0
